added missing endpoints

+++ app/service/Plants.py
@@ -54,23 +54,28 @@ class PlantsService():
             raise err
 
     async def get_log(self, log_id: int, token: str) -> LogSchema:
-        log: Log = self.plants_repository.get_log(log_id)
-        # TODO: Este print hace que los logs se parsen bien a LogSchemas.
-        # No quitar a menos que se encuentre una mejor solucion.
-        print(log)
         user_id = await UserService.get_user_id(token)
+        log: Log = self.plants_repository.get_log(log_id)
         plant = self.plants_repository.get_plant_by_id(log.plant_id)
         if user_id != plant.id_user:
             raise UserUnauthorized
 
+        # TODO: Este print hace que los logs se parsen bien a LogSchemas.
+        # No quitar a menos que se encuentre una mejor solucion.
+        print(log)
         return LogSchema.model_validate(log.__dict__)
 
-    def get_logs_by_user(
+    async def get_logs_by_user(
         self,
         user_id: int,
         year: int,
-        month: Optional[int]
+        month: Optional[int],
+        token: str
     ) -> List[LogSchema]:
+        id_user = await UserService.get_user_id(token)
+        if id_user != user_id:
+            raise UserUnauthorized
+
         if month:
             left = date(year, month, 1)
             right = date(year+1, 1, 1) if month == 12 else date(year, month+1, 1)
@@ -89,12 +94,25 @@ class PlantsService():
             logs
         ))
 
-    def update_log(
+    async def update_log(
         self,
         log_id: str,
-        log_update_set: LogPartialUpdateSchema
+        log_update_set: LogPartialUpdateSchema,
+        token: str
     ) -> Optional[LogSchema]:
         try:
+            user_id = await UserService.get_user_id(token)
+            log: Log = self.plants_repository.get_log(log_id)
+            plant = self.plants_repository.get_plant_by_id(log.plant_id)
+
+            if user_id != plant.id_user:
+                raise UserUnauthorized
+
+            if log_update_set.plant_id != None:
+                new_plant = self.plants_repository.get_plant_by_id(log_update_set.plant_id)
+                if user_id != new_plant.id_user:
+                    raise UserUnauthorized
+            
             self.plants_repository.update_log(
                 log_id,
                 log_update_set.title,
@@ -111,12 +129,19 @@ class PlantsService():
             self.plants_repository.rollback()
             raise err
 
-    def add_photo(
+    async def add_photo(
         self,
         id_log: str,
-        photo_create_set: LogPhotoCreateSchema
+        photo_create_set: LogPhotoCreateSchema,
+        token: str
     ) -> LogSchema:
         try:
+            user_id = await UserService.get_user_id(token)
+            log: Log = self.plants_repository.get_log(id_log)
+            plant = self.plants_repository.get_plant_by_id(log.plant_id)
+            if user_id != plant.id_user:
+                raise UserUnauthorized
+
             self.plants_repository.add(
                 LogPhoto(photo_link=photo_create_set.photo_link, log_id=id_log)
             )
@@ -129,8 +154,14 @@ class PlantsService():
             self.plants_repository.rollback()
             raise err
 
-    def delete_photo(self, id_log: int, id_photo: int):
+    async def delete_photo(self, id_log: int, id_photo: int, token: str):
         try:
+            user_id = await UserService.get_user_id(token)
+            log: Log = self.plants_repository.get_log(id_log)
+            plant = self.plants_repository.get_plant_by_id(log.plant_id)
+            if user_id != plant.id_user:
+                raise UserUnauthorized
+
             result_rowcount = self.plants_repository.delete_photo_from_log(
                 id_log, id_photo
             )
@@ -185,9 +216,12 @@ class PlantsService():
             self.plants_repository.get_all_plants(limit)
         ))
 
-    def get_plants_by_user(
-        self, id_user: int, limit: int
+    async def get_plants_by_user(
+        self, id_user: int, limit: int, token: str
     ) -> List[PlantSchema]:
+        user_id = await UserService.get_user_id(token)
+        if user_id != id_user:
+            raise UserUnauthorized
         return list(map(
             lambda pl: PlantSchema.model_validate(pl.__dict__),
             self.plants_repository.get_all_plants_by_user(id_user, limit)
